pertama

+++ routes/web.php
@@ -3,7 +3,8 @@
 use App\Http\Controllers\HeroController;
 use App\Http\Controllers\PromotionController;
 use App\Http\Controllers\ServiceController;
-use App\Models\Promotion;
+use App\Http\Controllers\PartnerController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -16,32 +17,49 @@ use Illuminate\Support\Facades\Route;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
+Route::get('/register',[UserController::class,'createRegister']);
+Route::post('/register',[UserController::class,'storeRegister']);
+
+Route::get('/login',[UserController::class,'login'])->name('login');
+Route::post('/login',[UserController::class,'proseslogin']);
 
 Route::get('/dashboard', function () {
     return view('pages.dashboard');
-});
+})->middleware('auth', 'checkRole:admin');
+//->middleware('auth'); digunakan untuk previkasi penguna 
 //pages hero//
-Route::get('/hero',[HeroController::class, 'index']); // 
-Route::get('/hero/create', [HeroController::class,'create']);
-Route::post('/hero',[HeroController::class,'store']);
-Route::get('hero/{hero}/edit', [HeroController::class, 'edit']);
-Route::put('/hero/{hero}', [HeroController::class,'update']);
-Route::delete('hero/{hero}',[HeroController::class,'destroy']);
+Route::get('/hero',[HeroController::class, 'index'])->middleware('auth', 'checkRole:admin');
+Route::get('/hero/create', [HeroController::class,'create'])->middleware('auth', 'checkRole:admin');
+Route::post('/hero',[HeroController::class,'store'])->middleware('auth', 'checkRole:admin');
+Route::get('hero/{hero}/edit', [HeroController::class, 'edit'])->middleware('auth', 'checkRole:admin');
+Route::put('/hero/{hero}', [HeroController::class,'update'])->middleware('auth', 'checkRole:admin');
+Route::delete('hero/{hero}',[HeroController::class,'destroy'])->middleware('auth', 'checkRole:admin');
 //akhir hero//
 
 // pages promotion//
-Route::get('/promotion',[PromotionController::class, 'index']);
-Route::get('promotion/create',[PromotionController::class, 'create']);
-Route::post('/promotion',[PromotionController::class, 'store']);
-Route::get('promotion/{promotion}/edit', [PromotionController::class, 'edit']);
-Route::put('/promotion/{promotion}', [PromotionController::class, 'update']);
-Route::delete('/promotion/{promotion}',[PromotionController::class, 'destroy']);
+Route::get('/promotion',[PromotionController::class, 'index'])->middleware('auth', 'checkRole:admin');
+Route::get('promotion/create',[PromotionController::class, 'create'])->middleware('auth', 'checkRole:admin');
+Route::post('/promotion',[PromotionController::class, 'store'])->middleware('auth', 'checkRole:admin');
+Route::get('promotion/{promotion}/edit', [PromotionController::class, 'edit'])->middleware('auth', 'checkRole:admin');
+Route::put('/promotion/{promotion}', [PromotionController::class, 'update'])->middleware('auth', 'checkRole:admin');
+Route::delete('/promotion/{promotion}',[PromotionController::class, 'destroy'])->middleware('auth', 'checkRole:admin');
 // akhir pages promotion//
 
 // pages service
-Route:: get('/service',[ServiceController::class, 'index']);
-Route:: get('/service/create',[ServiceController::class, 'create']);
+Route:: get('/service',[ServiceController::class, 'index'])->middleware('auth', 'checkRole:admin');
+Route:: get('/service/create',[ServiceController::class, 'create'])->middleware('auth', 'checkRole:admin');
 Route::post('/service',[ServiceController::class, 'store']);
-Route::get('service/{service}/edit', [ServiceController::class, 'edit']);
-Route::put('/service/{service}', [ServiceController::class, 'update']);
-Route::delete('/service/{service}',[ServiceController::class, 'destroy']);
+Route::get('service/{service}/edit', [ServiceController::class, 'edit'])->middleware('auth', 'checkRole:admin');
+Route::put('/service/{service}', [ServiceController::class, 'update'])->middleware('auth', 'checkRole:admin');
+Route::delete('/service/{service}',[ServiceController::class, 'destroy'])->middleware('auth', 'checkRole:admin');
+
+
+//Page Partner
+Route::get('/partner', [PartnerController::class, 'index'])->middleware('auth', 'checkRole:admin');
+Route::get('/partner/create', [PartnerController::class, 'create'])->middleware('auth', 'checkRole:admin');
+Route::post('partner', [PartnerController::class, 'store'])->middleware('auth', 'checkRole:admin');
+Route::get('/partner/{partner}/edit', [PartnerController::class, 'edit'])->middleware('auth', 'checkRole:admin');
+Route::put('/partner/{partner}', [PartnerController::class, 'update'])->middleware('auth', 'checkRole:admin');
+Route::delete('/partner/{partner}', [PartnerController::class, 'destroy'])->middleware('auth', 'checkRole:admin');
+
+Route::get('/logout',[UserController::class, 'logout'])->name('logout')->middleware('auth', 'checkRole:admin');
